+++ src/main/java/insane96mcp/shieldsplus/mixin/PlayerMixin.java
@@ -1,6 +1,6 @@
 package insane96mcp.shieldsplus.mixin;
 
-import insane96mcp.shieldsplus.module.base.BaseFeature;
+import insane96mcp.shieldsplus.module.BaseFeature;
 import insane96mcp.shieldsplus.setup.SPEnchantments;
 import insane96mcp.shieldsplus.world.item.enchantment.FastRecoveryEnchantment;
 import net.minecraft.world.entity.EntityType;