+++ src/main/java/insane96mcp/shieldsplus/module/BaseFeature.java
@@ -1,4 +1,4 @@
-package insane96mcp.shieldsplus.module.base;
+package insane96mcp.shieldsplus.module;
 
 import insane96mcp.insanelib.base.Feature;
 import insane96mcp.insanelib.base.Label;
@@ -6,20 +6,27 @@ import insane96mcp.insanelib.base.Module;
 import insane96mcp.insanelib.base.config.Config;
 import insane96mcp.insanelib.base.config.LoadFeature;
 import insane96mcp.shieldsplus.ShieldsPlus;
+import insane96mcp.shieldsplus.data.ShieldDefinition;
+import insane96mcp.shieldsplus.data.ShieldDefinitionReloader;
 import insane96mcp.shieldsplus.setup.SPEnchantments;
 import insane96mcp.shieldsplus.world.item.SPShieldItem;
 import insane96mcp.shieldsplus.world.item.enchantment.*;
 import net.minecraft.world.item.ShieldItem;
+import net.minecraftforge.api.distmarker.Dist;
+import net.minecraftforge.api.distmarker.OnlyIn;
 import net.minecraftforge.event.TickEvent;
 import net.minecraftforge.event.entity.living.LivingHurtEvent;
 import net.minecraftforge.event.entity.living.ShieldBlockEvent;
+import net.minecraftforge.event.entity.player.ItemTooltipEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
+
+import java.util.Optional;
 
 @Label(name = "Shields+")
 @LoadFeature(module = ShieldsPlus.RESOURCE_PREFIX + "base", canBeDisabled = false)
 public class BaseFeature extends Feature {
     @Config(min = 0)
-    @Label(name = "Shield Windup", description = "In vanilla when you start blocking with a shield, there's a 0.25 seconds (5 ticks) window where you are still not blocking. By default the windup is disabled.")
+    @Label(name = "Shield Windup", description = "In vanilla when you start blocking with a shield, there's a 0.25 seconds (5 ticks) window where you are still not blocking. By default the windup is removed.")
     public static Integer shieldWindup = 0;
     @Config
     @Label(name = "Shields Block Fixed Damage Amount", description = "If true shields will block only a certain amount of damage. If false the vanilla behaviour is used.")
@@ -40,10 +47,19 @@ public class BaseFeature extends Feature {
 
     @SubscribeEvent
     public void onShieldBlock(ShieldBlockEvent event) {
-        if (shieldBlockFixedDamageAmount && event.getEntity().getUseItem().getItem() instanceof SPShieldItem) {
-            double baseBlockedDamage = ((SPShieldItem) event.getEntity().getUseItem().getItem()).getBlockedDamage(event.getEntity().getUseItem(), event.getEntity(), event.getEntity().level());
-            float blockedDamage = (float) (baseBlockedDamage + ReinforcedEnchantment.getDamageBlocked(event.getEntity().getUseItem()));
-            event.setBlockedDamage(blockedDamage);
+        if (shieldBlockFixedDamageAmount) {
+            if (event.getEntity().getUseItem().getItem() instanceof SPShieldItem spShieldItem) {
+                double baseBlockedDamage = spShieldItem.getBlockedDamage(event.getEntity().getUseItem(), event.getEntity(), event.getEntity().level());
+                float blockedDamage = (float) (baseBlockedDamage + ReinforcedEnchantment.getDamageBlocked(event.getEntity().getUseItem()));
+                event.setBlockedDamage(blockedDamage);
+            }
+            else {
+                Optional<ShieldDefinition> shieldDefinition = ShieldDefinitionReloader.getShieldDefinition(event.getEntity().getUseItem());
+                shieldDefinition.ifPresent(def -> {
+                    float blockedDamage = (def.blockedDamage + ReinforcedEnchantment.getDamageBlocked(event.getEntity().getUseItem()));
+                    event.setBlockedDamage(blockedDamage);
+                });
+            }
         }
 
         //Process blocking enchantments
@@ -72,6 +88,25 @@ public class BaseFeature extends Feature {
 
         LightweightEnchantment.onTick(event.player);
         ShieldBashEnchantment.onTick(event.player);
+    }
+
+    @OnlyIn(Dist.CLIENT)
+    @SubscribeEvent
+    public void onTooltip(ItemTooltipEvent event) {
+        if (!this.isEnabled()
+                || !shieldBlockFixedDamageAmount
+                || event.getItemStack().getItem() instanceof SPShieldItem)
+            return;
+
+        Optional<ShieldDefinition> shieldDefinition = ShieldDefinitionReloader.getShieldDefinition(event.getItemStack());
+        if (shieldDefinition.isEmpty())
+            return;
+        SPShieldItem.addDamageBlockedText(event.getItemStack(), event.getToolTip(), shieldDefinition.get().blockedDamage);
+        /*int useDuration = this.getUseDuration(itemStack);
+        if (useDuration < 72000) {
+            components.add(Component.translatable(BLOCKING_TIME, ShieldsPlus.ONE_DECIMAL_FORMATTER.format(useDuration / 20f)).withStyle(ChatFormatting.BLUE));
+            components.add(Component.translatable(COOLDOWN, ShieldsPlus.ONE_DECIMAL_FORMATTER.format(this.getCooldown(itemStack, null, level) / 20f)).withStyle(ChatFormatting.BLUE));
+        }*/
     }
 
     public static boolean shouldRemoveShieldWindup() {